requirements.txt
aiofiles==23.1.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.11.0
anyio==4.4.0
appscript==1.2.5
asgiref==3.8.1
async-timeout==3.0.1
asyncpg==0.28.0
attrs==24.2.0
beautifulsoup4==4.12.2
certifi==2024.7.4
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
databases==0.7.0
fastapi==0.112.0
frozenlist==1.4.1
h11==0.14.0
idna==3.7
lxml==5.3.0
Mako==1.3.5
MarkupSafe==2.1.5
multidict==6.0.5
psutil==6.0.0
psycopg2-binary==2.9.9
pydantic==1.10.17
python-dotenv==1.0.0
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.53
starlette==0.37.2
typing_extensions==4.12.2
urllib3==2.2.2
uvicorn==0.27.1
vercel==0.2.1
xlwings==0.30.10
yarl==1.9.4
